services/api/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.30.1
sqlmodel==0.0.16
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic-settings==2.2.1
alembic==1.13.1
redis==5.0.3
httpx==0.27.0
loguru==0.7.2
python-slugify==8.0.4
email-validator==2.1.1
